snowpark/requirements.txt
pandas==2.2.1
pyarrow==15.0.2
snowflake-snowpark-python==1.14.0
